+++ src/main/java/com/global/api/entities/enums/TransactionType.java
@@ -84,6 +84,7 @@ public enum TransactionType implements IFlag {
     GetAccountDetails,
     GetAccountBalance,
     GetTokenInfo,
+    DeleteOpenTab,
     CheckQueryInfo,
     OrderDevice,
     SurchargeEligibilityLookup;