+++ src/main/java/com/global/api/terminals/upa/Entities/Enums/UpaMessageId.java
@@ -25,6 +25,6 @@ public enum UpaMessageId {
     GetSignature,
     DeleteSAF,
     RegisterPOS,
-
+    DeletePreAuth,
     PrintData
 }